+++ yash-builtin/src/typeset/syntax.rs
@@ -251,9 +251,6 @@ fn try_parse_long<'a, I: Iterator<Item = Field>>(
     };
 
     let (name, negate) = if let Some(name) = field.value.strip_prefix("--") {
-        if name.is_empty() {
-            return Ok(None);
-        }
         (name, false)
     } else if let Some(name) = field.value.strip_prefix("++") {
         (name, true)
@@ -297,6 +294,9 @@ pub fn parse<'a>(
     let mut args = args.into_iter().peekable();
     let mut options = Vec::new();
     loop {
+        if args.next_if(|arg| arg.value == "--").is_some() {
+            break;
+        }
         if try_parse_short(option_specs, &mut args, &mut options)? {
             continue;
         }
@@ -459,6 +459,24 @@ mod tests {
     }
 
     #[test]
+    fn parse_many_short_options() {
+        let args = Field::dummies(["-p", "+xr"]);
+        let result = parse(ALL_OPTIONS, args.clone()).unwrap();
+        assert_matches!(&result.0[..], [option1, option2, option3] => {
+            assert_eq!(option1.spec, &PRINT_OPTION);
+            assert_eq!(option1.state, State::On);
+            assert_eq!(option1.location, Location::dummy("-p"));
+            assert_eq!(option2.spec, &EXPORT_OPTION);
+            assert_eq!(option2.state, State::Off);
+            assert_eq!(option2.location, Location::dummy("+xr"));
+            assert_eq!(option3.spec, &READONLY_OPTION);
+            assert_eq!(option3.state, State::Off);
+            assert_eq!(option3.location, Location::dummy("+xr"));
+        });
+        assert_eq!(result.1, []);
+    }
+
+    #[test]
     fn parse_long_print_option_without_operands() {
         let result = parse(ALL_OPTIONS, Field::dummies(["--print"])).unwrap();
         assert_matches!(&result.0[..], [option] => {
@@ -514,6 +532,18 @@ mod tests {
             assert_eq!(option.location, Location::dummy("++export"));
         });
         assert_eq!(result.1, []);
+    }
+
+    #[test]
+    fn parse_separator() {
+        let args = Field::dummies(["-p", "--", "-x"]);
+        let result = parse(ALL_OPTIONS, args.clone()).unwrap();
+        assert_matches!(&result.0[..], [option] => {
+            assert_eq!(option.spec, &PRINT_OPTION);
+            assert_eq!(option.state, State::On);
+            assert_eq!(option.location, Location::dummy("-p"));
+        });
+        assert_eq!(result.1, Field::dummies(["-x"]));
     }
 
     #[test]