+++ yash-cli/tests/scripted_test/shift-p.sh
@@ -84,3 +84,9 @@ __IN__
 [2][y  y][z]
 [3][a][b  b][c]
 __OUT__
+
+test_oE -e 0 'separator preceding operand' -es a b c d e
+shift -- 2 && bracket "$#" "$@"
+__IN__
+[3][c][d][e]
+__OUT__